backend/requirements.txt
# Django core
Django==4.2.8
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
setuptools>=65.0.0

# Database
psycopg2-binary==2.9.9

# Redis & Celery
redis==5.0.1
celery==5.3.4
django-celery-beat==2.5.0

# CORS
django-cors-headers==4.3.1

# Filtering
django-filter==23.5

# PDF generation
weasyprint==59.0
pydyf==0.10.0

# Environment variables
python-dotenv==1.0.0

# Date handling
python-dateutil==2.8.2

# Utilities
Pillow==10.2.0

# Production WSGI Server
gunicorn==21.2.0
